// apps/previ/src/protocols/v1/acknowledge.schema.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface AcknowledgeSchema {
  /**
   * The version of the protocol.
   */
  protocolVersion: string;
  /**
   * The type of the message.
   */
  messageType: "acknowledge";
  /**
   * A human-readable message indicating successful initialization.
   */
  message: string;
}