// apps/previ/src/protocols/v1/init_progress.schema.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface InitProgressSchema {
  /**
   * The version of the protocol.
   */
  protocolVersion: string;
  /**
   * The type of the message.
   */
  messageType: "init_progress";
  /**
   * A human-readable message indicating the progress of the preview process.
   */
  message: string;
  /**
   * A number between 0 and 1 indicating the progress (0 = start, 1 = complete).
   */
  progress: number;
}